// src/sdks/footprintV2ApiDocs.ts
export interface ProductionMaterials {
  id: number;
  productionBusinessId: number;
  backingMaterial?: string;
  createBy?: string;
  createTime?: Date;
  delFlag?: number;
  discharge?: string;
  distance?: number;
  factorId?: number;
  factorName?: string;
  factorSource?: string;
  factorType?: number;
  factorUnit?: string;
  factorValue?: string;
  factorYear?: string;
  formula?: string;
  maMeasure?: string;
  materialName?: string;
  materialsType?: string;
  measureCode?: string;
  orgId?: number;
  percentMeasure?: string;
  processModelId?: number;
  processModelParentId?: number;
  status?: number;
  type?: number;
  unitCode?: string;
  updateBy?: string;
  updateTime?: Date;
  weight?: number;
}